version 1.0.10:
issue #226: develop minimal logic (M0).

lock mechanism for axiom-inclusion and axiom-interpretation.

issue #229: elaborate intuitionistic logic (j0) theory.

issue #231: compose the theory declaration and list the extended theories.

issue #232: provide the capability to forbid the inclusion of new inference-rules in a theory.

issue #233: provide the capability to forbid the inclusion of new axioms / definitions in a theory.